PaymentService/pyproject.toml
[tool.poetry]
name = "paymentservice"
version = "0.1.0"
description = ""
authors = ["JonAJ <jonajmail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["all"], version = "^0.32.1"}
ipython = "^8.29.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
punq = "^0.7.0"
httpx = "^0.27.2"
faker = "^33.0.0"
pydantic-settings = "^2.6.1"
python-multipart = "^0.0.17"
asyncpg = "^0.30.0"
django-environ = "^0.11.2"
nest-asyncio = "^1.6.0"
YooMoney = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
